Extract Explain::ArchScope for reuse
Factor the per-arch leaf scoping (arch_values, sections, other_leaves,
default_label) and the shared PathFacts cache out of Summary into a small
ArchScope class, so the upcoming audit command can split a filter by
architecture the same way explain does. Summary delegates; output unchanged.
